driver=nl80211 does not work on Pi Zero W
I tried following the instructions on a Pi Zero W and could not get it to work.  A Linux sysadmin at work helped me track down the issue to the driver=nl80211 line as there was an error saying driver not found.  When we removed that line it worked perfectly.